+++ parts/dashboardConfig.js
