+++ parts/newDocumentStructure.js
