+++ parts/publishPluginAction.js
