+++ parts/resolveDocumentActions.js
